refactored onyx.py to move logic related to drop_counter to separate class
